# AppFormationFrontend/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": { "title": "OpenAPI definition", "version": "v0" },
  "servers": [
    {
      "url": "http://localhost:8081/api/v1",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/auth/uploadFile/{email}": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "uploadFile",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": ["file"],
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegistrationRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/auth/registerParticipant": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "registerParticipant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegistrationRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/auth/registerInstructor": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "registerInstructor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegistrationRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/auth/registerAdmin": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "registerAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegistrationRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "logout",
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/auth/authenticate": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "authenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthenticationRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user/all": {
      "get": {
        "tags": ["User"],
        "operationId": "findAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          }
        }
      }
    },
    "/auth/activate-account": {
      "get": {
        "tags": ["Authentication"],
        "operationId": "confirm",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    }
  },
  "components": {
    "schemas": {
      "RegistrationRequest": {
        "required": [
          "cin",
          "email",
          "firstname",
          "lastname",
          "password",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "firstname": { "type": "string" },
          "lastname": { "type": "string" },
          "email": { "type": "string" },
          "password": {
            "maxLength": 2147483647,
            "minLength": 8,
            "type": "string"
          },
          "cin": { "maxLength": 8, "minLength": 8, "type": "string" },
          "dateOfBirth": { "type": "string", "format": "date" },
          "phoneNumber": { "type": "string" },
          "speciality": { "type": "string" },
          "workplace": { "type": "string" },
          "company": { "type": "string" },
          "pdfFile": { "type": "string" }
        }
      },
      "AuthenticationRequest": {
        "required": ["email", "password"],
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "password": {
            "maxLength": 2147483647,
            "minLength": 8,
            "type": "string"
          }
        }
      },
      "AuthenticationResponse": {
        "type": "object",
        "properties": { "token": { "type": "string" } }
      },
      "GrantedAuthority": {
        "type": "object",
        "properties": { "authority": { "type": "string" } }
      },
      "Role": {
        "type": "object",
        "properties": {
          "roleId": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "createdDate": { "type": "string", "format": "date-time" },
          "lastModifiedDate": { "type": "string", "format": "date-time" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "userId": { "type": "integer", "format": "int64" },
          "cin": { "type": "string" },
          "firstname": { "type": "string" },
          "lastname": { "type": "string" },
          "dateOfBirth": { "type": "string", "format": "date" },
          "email": { "type": "string" },
          "password": { "type": "string" },
          "accountLocked": { "type": "boolean" },
          "enabled": { "type": "boolean" },
          "verified": { "type": "boolean" },
          "phoneNumber": { "type": "string" },
          "speciality": { "type": "string" },
          "workplace": { "type": "string" },
          "company": { "type": "string" },
          "pdfFile": { "type": "string" },
          "createdDate": { "type": "string", "format": "date-time" },
          "lastModifiedDate": { "type": "string", "format": "date-time" },
          "roles": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Role" }
          },
          "name": { "type": "string" },
          "username": { "type": "string" },
          "authorities": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/GrantedAuthority" }
          },
          "accountNonLocked": { "type": "boolean" },
          "accountNonExpired": { "type": "boolean" },
          "credentialsNonExpired": { "type": "boolean" }
        }
      }
    }
  }
}